Renamed matterhorn package
