+++ java/matterhorn_data_tools/src/main/java/edu/harvard/data/matterhorn/cli/Command.java
@@ -1,4 +1,4 @@
-package edu.harvard.data.canvas.cli;
+package edu.harvard.data.matterhorn.cli;
 
 import java.io.IOException;
 import java.util.concurrent.ExecutorService;
