+++ java/matterhorn_data_tools/src/main/java/edu/harvard/data/matterhorn/cli/PreVerifyCommand.java
@@ -1,4 +1,4 @@
-package edu.harvard.data.canvas.cli;
+package edu.harvard.data.matterhorn.cli;
 
 import java.io.IOException;
 import java.net.URI;
@@ -10,20 +10,18 @@ import org.apache.logging.log4j.Logger;
 import org.kohsuke.args4j.Argument;
 import org.kohsuke.args4j.Option;
 
-import edu.harvard.data.AwsUtils;
 import edu.harvard.data.DataConfiguration;
 import edu.harvard.data.DataConfigurationException;
-import edu.harvard.data.FormatLibrary;
 import edu.harvard.data.ReturnStatus;
 import edu.harvard.data.VerificationException;
 import edu.harvard.data.Verifier;
-import edu.harvard.data.canvas.phase_0.Phase0PostVerifier;
-import edu.harvard.data.canvas.phase_1.Phase1PostVerifier;
-import edu.harvard.data.canvas.phase_2.Phase2PostVerifier;
-import edu.harvard.data.canvas.phase_3.Phase3PostVerifier;
+import edu.harvard.data.matterhorn.phase_0.Phase0PreVerifier;
+import edu.harvard.data.matterhorn.phase_1.Phase1PreVerifier;
+import edu.harvard.data.matterhorn.phase_2.Phase2PreVerifier;
+import edu.harvard.data.matterhorn.phase_3.Phase3PreVerifier;
 import edu.harvard.data.schema.UnexpectedApiResponseException;
 
-public class PostVerifyCommand implements Command {
+public class PreVerifyCommand implements Command {
 
   private static final Logger log = LogManager.getLogger();
 
@@ -33,14 +31,11 @@ public class PostVerifyCommand implements Command {
   @Option(name = "-i", usage = "UUID for the dump, generated by the Canvas Data API. Required for phase 0.", metaVar = "uuid", required = false)
   public String dumpId;
 
-  @Option(name = "-in", usage = "Location of original data files. Required for phases > 0.", metaVar = "/path/to/data", required = false)
+  @Option(name = "-in", usage = "Location of data files from the previous phase. Required for phases > 0.", metaVar = "/path/to/data", required = false)
   public String inputDir;
 
-  @Option(name = "-out", usage = "Location of processed data files. Required for phases > 0.", metaVar = "/path/to/data", required = false)
+  @Option(name = "-out", usage = "Temporary location to store verification data. Required for phases > 0.", metaVar = "/path/to/temp", required = false)
   public String outputDir;
-
-  @Option(name = "-verify", usage = "Location of verification files. Required for phases > 0.", metaVar = "/path/to/verify", required = false)
-  public String verifyDir;
 
   @Override
   public ReturnStatus execute(final DataConfiguration config, final ExecutorService exec)
@@ -67,48 +62,47 @@ public class PostVerifyCommand implements Command {
     } catch (final URISyntaxException e) {
       throw new DataConfigurationException(e);
     }
-    final AwsUtils aws = new AwsUtils();
-    final FormatLibrary formats = new FormatLibrary();
     switch (phase) {
     case 0:
-      return new Phase0PostVerifier(dumpId, aws, config.getScratchDir(), exec);
+      return new Phase0PreVerifier();
     case 1:
-      return new Phase1PostVerifier(hdfsService, inputDir, outputDir, verifyDir);
+      return new Phase1PreVerifier(config, hdfsService, inputDir, outputDir);
     case 2:
-      return new Phase2PostVerifier();
+      return new Phase2PreVerifier();
     case 3:
-      return new Phase3PostVerifier();
+      return new Phase3PreVerifier();
     }
     throw new ArgumentError("Invalid phase " + phase);
   }
 
-  private boolean checkArguments() throws DataConfigurationException {
+  private boolean checkArguments() {
     if (phase == 0) {
       if (dumpId == null) {
         log.error("Dump ID is required for Phase 0");
         return false;
       }
     } else {
-      outputDir = checkPath(outputDir);
-      inputDir = checkPath(inputDir);
-      verifyDir = checkPath(verifyDir);
+      if (inputDir == null) {
+        log.error("Input directory is required for all phases other than zero.");
+        return false;
+      }
+      if (outputDir == null) {
+        log.error("Output directory is required for all phases other than zero.");
+        return false;
+      }
+      if (!inputDir.toLowerCase().startsWith("hdfs://")) {
+        inputDir = "hdfs://" + inputDir;
+      }
+      if (!outputDir.toLowerCase().startsWith("hdfs://")) {
+        outputDir = "hdfs://" + outputDir;
+      }
     }
     return true;
   }
 
-  private String checkPath(final String dir) throws DataConfigurationException {
-    if (dir == null) {
-      throw new DataConfigurationException("Input directory is required for all phases other than zero.");
-    }
-    if (!dir.toLowerCase().startsWith("hdfs://")) {
-      return "hdfs://" + dir;
-    }
-    return dir;
-  }
-
   @Override
   public String getDescription() {
-    return "Perform verification operations following a phase.";
+    return "Perform any initial operations to set up verification before a phase.";
   }
 
 }
