+++ java/matterhorn_data_tools/src/main/java/edu/harvard/data/matterhorn/phase_0/Phase0PostVerifier.java
@@ -1,4 +1,4 @@
-package edu.harvard.data.canvas.phase_0;
+package edu.harvard.data.matterhorn.phase_0;
 
 import java.io.File;
 import java.io.IOException;
