+++ java/matterhorn_data_tools/src/main/java/edu/harvard/data/matterhorn/phase_0/Phase0PreVerifier.java
@@ -1,4 +1,4 @@
-package edu.harvard.data.canvas.phase_3;
+package edu.harvard.data.matterhorn.phase_0;
 
 import java.io.IOException;
 
@@ -8,12 +8,12 @@ import org.apache.logging.log4j.Logger;
 import edu.harvard.data.VerificationException;
 import edu.harvard.data.Verifier;
 
-public class Phase3PostVerifier implements Verifier {
+public class Phase0PreVerifier implements Verifier {
   private static final Logger log = LogManager.getLogger();
 
   @Override
   public void verify() throws VerificationException, IOException {
-    log.info("Running post-verifier for phase 1");
+    log.info("Running pre-verifier for phase 0");
   }
 
 }
