+++ java/matterhorn_data_tools/src/main/java/edu/harvard/data/matterhorn/phase_2/Phase2PostVerifier.java
@@ -1,4 +1,4 @@
-package edu.harvard.data.canvas.phase_2;
+package edu.harvard.data.matterhorn.phase_2;
 
 import java.io.IOException;
