+++ java/matterhorn_data_tools/src/main/java/edu/harvard/data/matterhorn/phase_3/Phase3PostVerifier.java
@@ -1,4 +1,4 @@
-package edu.harvard.data.canvas.phase_3;
+package edu.harvard.data.matterhorn.phase_3;
 
 import java.io.IOException;
 
@@ -8,12 +8,12 @@ import org.apache.logging.log4j.Logger;
 import edu.harvard.data.VerificationException;
 import edu.harvard.data.Verifier;
 
-public class Phase3PreVerifier implements Verifier {
+public class Phase3PostVerifier implements Verifier {
   private static final Logger log = LogManager.getLogger();
 
   @Override
   public void verify() throws VerificationException, IOException {
-    log.info("Running pre-verifier for phase 3");
+    log.info("Running post-verifier for phase 1");
   }
 
 }
